+++ CHANGELOG.md
@@ -53,6 +53,17 @@ from `algosdk` and sends it to the network.
 - Fix `TxParams.noteb64` encoding - should use base64 decoder rather than TextEncoder.
 - Fix `ed25519verify` opcode implementation. Previously the signature was only checked against the data not the concatenation of "ProgData"||program||data. Additionally test scenarios was added to check the correct implementation. 
 
+### Breaking Changes
+
+#### @algo-builder/runtime
+
+- `executeTx` now returns [TxnReceipt](https://github.com/scale-it/algo-builder/blob/master/packages/runtime/src/types.ts#L411) instead of `TxReceipt[]`.
+- `fundLsig` now returns [TxnReceipt](https://github.com/scale-it/algo-builder/blob/master/packages/runtime/src/types.ts#L411) instead of `TxReceipt[]`.
+
+#### @algo-builder/web
+
+- `executeTx` promise now returns [TxnReceipt](https://github.com/scale-it/algo-builder/blob/master/packages/web/src/types.ts#L458) instead of `algosdk.modelsv2.PendingTransactionResponse`.
+
 ### Examples
 
 #### DAO